Fix responsiveness
Add max-width to a container for the mondrian-grid of 800px to keep it in the 10x10 matrix.
Moreover, I had to rearrange some elements of the grid so that it would resemble the original paiting. For example child 16 had to be moved to child 19 in order for the pattern to work.